Fixed. ProgressBarの表示
