+++ Task36/app/src/main/AndroidManifest.xml
@@ -22,16 +22,6 @@
         </activity>
         <activity android:name=".PostImage" />
 
-        <receiver
-            android:name=".MyResultReceiver"
-            android:exported="false">
-            <intent-filter>
-                <action android:name="com.twitter.sdk.android.tweetcomposer.UPLOAD_SUCCESS"/>
-                <action android:name="com.twitter.sdk.android.tweetcomposer.UPLOAD_FAILURE"/>
-                <action android:name="com.twitter.sdk.android.tweetcomposer.TWEET_COMPOSE_CANCEL"/>
-            </intent-filter>
-        </receiver>
-
     </application>
